# stroom-auth-service-api/src/main/java/stroom/auth/service/api/model/IdTokenRequest.java
/*
 * Stroom Auth API
 * Various APIs for interacting with authentication, users, and tokens.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package stroom.auth.service.api.model;

import java.util.Objects;
import com.google.gson.annotations.SerializedName;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;

/**
 * A request to retrieve an ID token using an access code.
 */
@ApiModel(description = "A request to retrieve an ID token using an access code.")
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2017-11-01T10:45:36.027Z")
public class IdTokenRequest {
  @SerializedName("sessionId")
  private String sessionId = null;

  @SerializedName("accessCode")
  private String accessCode = null;

  @SerializedName("requestingClientId")
  private String requestingClientId = null;

  public IdTokenRequest sessionId(String sessionId) {
    this.sessionId = sessionId;
    return this;
  }

   /**
   * The session ID to which this request is linked.
   * @return sessionId
  **/
  @ApiModelProperty(example = "null", required = true, value = "The session ID to which this request is linked.")
  public String getSessionId() {
    return sessionId;
  }

  public void setSessionId(String sessionId) {
    this.sessionId = sessionId;
  }

  public IdTokenRequest accessCode(String accessCode) {
    this.accessCode = accessCode;
    return this;
  }

   /**
   * The previously provided access code.
   * @return accessCode
  **/
  @ApiModelProperty(example = "null", required = true, value = "The previously provided access code.")
  public String getAccessCode() {
    return accessCode;
  }

  public void setAccessCode(String accessCode) {
    this.accessCode = accessCode;
  }

  public IdTokenRequest requestingClientId(String requestingClientId) {
    this.requestingClientId = requestingClientId;
    return this;
  }

   /**
   * The clientId of the relying party requesting the id_token.
   * @return requestingClientId
  **/
  @ApiModelProperty(example = "null", required = true, value = "The clientId of the relying party requesting the id_token.")
  public String getRequestingClientId() {
    return requestingClientId;
  }

  public void setRequestingClientId(String requestingClientId) {
    this.requestingClientId = requestingClientId;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    IdTokenRequest idTokenRequest = (IdTokenRequest) o;
    return Objects.equals(this.sessionId, idTokenRequest.sessionId) &&
        Objects.equals(this.accessCode, idTokenRequest.accessCode) &&
        Objects.equals(this.requestingClientId, idTokenRequest.requestingClientId);
  }

  @Override
  public int hashCode() {
    return Objects.hash(sessionId, accessCode, requestingClientId);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class IdTokenRequest {\n");
    
    sb.append("    sessionId: ").append(toIndentedString(sessionId)).append("\n");
    sb.append("    accessCode: ").append(toIndentedString(accessCode)).append("\n");
    sb.append("    requestingClientId: ").append(toIndentedString(requestingClientId)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
  
}