# swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Various APIs for interacting with authentication, users, and tokens."
  version: "v1"
  title: "Stroom Auth API"
  contact:
    name: "GCHQ"
    url: "https://github.com/gchq/stroom-auth"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
...
  /authentication/v1/authenticate:
    get:
      tags:
      - "Authentication"
      summary: "Submit an OpenId AuthenticationRequest."
      description: ""
      operationId: "handleAuthenticationRequest"
      produces:
      - "application/json"
      parameters:
      - name: "scope"
        in: "query"
        required: false
        type: "string"
      - name: "response_type"
        in: "query"
        required: false
        type: "string"
      - name: "client_id"
        in: "query"
        required: true
        type: "string"
      - name: "redirect_url"
        in: "query"
        required: true
        type: "string"
      - name: "nonce"
        in: "query"
        required: true
        type: "string"
      - name: "state"
        in: "query"
        required: false
        type: "string"
      - name: "sessionId"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
    post:
      tags:
      - "Authentication"
      summary: "Handle a login request made using username and password credentials."
      description: ""
      operationId: "handleLogin"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Credentials"
        required: false
        schema:
          $ref: "#/definitions/Credentials"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
  /authentication/v1/idToken:
    get:
      tags:
      - "Authentication"
      summary: "Convert a previously provided access code into an ID token"
      description: ""
      operationId: "getIdTokenWithGet"
      produces:
      - "application/json"
      parameters:
      - name: "accessCode"
        in: "query"
        required: true
        type: "string"
      - name: "clientId"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
    post:
      tags:
      - "Authentication"
      summary: "Convert a previously provided access code into an ID token"
      description: ""
      operationId: "getIdTokenWithPost"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "idTokenRequest"
        required: false
        schema:
          $ref: "#/definitions/IdTokenRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
  /authentication/v1/reset/{email}:
    get:
      tags:
      - "Authentication"
      summary: "Reset a user account using an email address."
      description: ""
      operationId: "resetEmail"
      produces:
      - "application/json"
      parameters:
      - name: "email"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
  /authentication/v1/verify/{token}:
    get:
      tags:
      - "Authentication"
      summary: "Verify the authenticity and current-ness of a JWS token."
      description: ""
      operationId: "verifyToken"
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
...
definitions:
...
  Credentials:
    type: "object"
    required:
    - "email"
    - "password"
    properties:
      sessionId:
        type: "string"
      requestingClientId:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
  IdTokenRequest:
    type: "object"
    required:
    - "accessCode"
    - "requestingClientId"
    - "sessionId"
    properties:
      sessionId:
        type: "string"
        description: "The session ID to which this request is linked."
      accessCode:
        type: "string"
        description: "The previously provided access code."
      requestingClientId:
        type: "string"
        description: "The clientId of the relying party requesting the id_token."
    description: "A request to retrieve an ID token using an access code."
...